FossilOrigin-Name: 270d2fb95ce3ba86fcf5427bcfaa463cdeb79a4e63784faf6f3e37eaf23bf564

+++ src/setup.c
@@ -166,7 +166,7 @@ void setup_ulist(void){
     style_header("User List");
     @ <table border=1 cellpadding=2 cellspacing=0 class='userTable'>
     @ <thead><tr>
-    @   <th>UID <th>Category
+    @   <th>Category
     @   <th>Capabilities (<a href='%R/setup_ucap_list'>key</a>)
     @   <th>Info <th>Last Change</tr></thead>
     @ <tbody>
@@ -182,7 +182,6 @@ void setup_ulist(void){
       const char *zCap = db_column_text(&s, 2);
       const char *zDate = db_column_text(&s, 4);
       @ <tr>
-      @ <td><a href='setup_uedit?id=%d(uid)'>%d(uid)</a>
       @ <td><a href='setup_uedit?id=%d(uid)'>%h(zLogin)</a>
       @ <td>%h(zCap)
 
@@ -210,12 +209,13 @@ void setup_ulist(void){
   }
   @ </tbody></table>
   @ <div class='section'>Users</div>
-  @ <table border=1 cellpadding=2 cellspacing=0 class='userTable' id='userlist'>
+  @ <table border=1 cellpadding=2 cellspacing=0 class='userTable sortable' \
+  @  data-column-types='ktxTTK' data-init-sort='2'>
   @ <thead><tr>
-  @ <th>ID<th>Login Name<th>Caps<th>Info<th>Date<th>Expire<th>Last Login</tr></thead>
+  @ <th>Login Name<th>Caps<th>Info<th>Date<th>Expire<th>Last Login</tr></thead>
   @ <tbody>
   db_multi_exec(
-     "CREATE TEMP TABLE lastAccess(uname TEXT PRIMARY KEY, atime REAL) WITHOUT ROWID;"
+    "CREATE TEMP TABLE lastAccess(uname TEXT PRIMARY KEY, atime REAL) WITHOUT ROWID;"
   );
   if( db_table_exists("repository","accesslog") ){
     db_multi_exec(
@@ -223,7 +223,8 @@ void setup_ulist(void){
       " SELECT uname, max(mtime) FROM ("
       "    SELECT uname, mtime FROM accesslog WHERE success"
       "    UNION ALL"
-      "    SELECT login AS uname, rcvfrom.mtime AS mtime FROM rcvfrom JOIN user USING(uid))"
+      "    SELECT login AS uname, rcvfrom.mtime AS mtime"
+      "      FROM rcvfrom JOIN user USING(uid))"
       " GROUP BY 1;"
     );
   }
@@ -257,7 +258,6 @@ void setup_ulist(void){
       zAge = human_readable_age(rNow - rATime);
     }
     @ <tr>
-    @ <td><a href='setup_uedit?id=%d(uid)'>%d(uid)</a>
     @ <td data-sortkey='%h(zSortKey)'><a href='setup_uedit?id=%d(uid)'>%h(zLogin)</a>
     @ <td>%h(zCap)
     @ <td>%h(zInfo)
@@ -269,7 +269,7 @@ void setup_ulist(void){
   }
   @ </tbody></table>
   db_finalize(&s);
-  output_table_sorting_javascript("userlist","nktxTTK",2);
+  style_table_sorter();
   style_footer();
 }
 
@@ -452,7 +452,7 @@ void user_edit(void){
   ** modified user record.  After writing the user record, redirect
   ** to the page that displays a list of users.
   */
-  doWrite = cgi_all("login","info","pw") && !higherUser;
+  doWrite = cgi_all("login","info","pw") && !higherUser && cgi_csrf_safe(1);
   if( doWrite ){
     char c;
     char zCap[50], zNm[4];
@@ -602,9 +602,10 @@ void user_edit(void){
 
   /* Begin generating the page
   */
-  style_submenu_element("Cancel", cgi_referer("setup_ulist"));
+  style_submenu_element("Cancel", "%s", cgi_referer("setup_ulist"));
   if( uid ){
     style_header("Edit User %h", zLogin);
+    style_submenu_element("Access Log", "%R/access_log?u=%t", zLogin);
   }else{
     style_header("Add A New User");
   }
@@ -617,43 +618,6 @@ void user_edit(void){
     @ <input type="hidden" name="pw" value="*">
   }
   @ <input type="hidden" name="referer" value="%h(cgi_referer("setup_ulist"))">
-  @ <script>
-  @ function updateCapabilityString(){
-  @   /*
-  @   ** This function updates the "#usetupEditCapability" span content
-  @   ** with the capabilities selected by the interactive user, based
-  @   ** upon the state of the capability checkboxes.
-  @   */
-  @   try {
-  @     var inputs = document.getElementsByTagName('input');
-  @     if( inputs && inputs.length ){
-  @       var output = document.getElementById('usetupEditCapability');
-  @       if( output ){
-  @         var permsIds = [], x = 0;
-  @         for(var i = 0; i < inputs.length; i++){
-  @           var e = inputs[i];
-  @           if( !e.name || !e.type ) continue;
-  @           if( e.type.toLowerCase()!=='checkbox' ) continue;
-  @           if( e.name.length===2 && e.name[0]==='a' ){
-  @             // looks like a capability checkbox
-  @             if( e.checked ){
-  @               // grab the second character of the element
-  @               // name, which is the textual flag for this
-  @               // capability, and then add it to the result
-  @               // array.
-  @               permsIds[x++] = e.name[1];
-  @             }
-  @           }
-  @         }
-  @         permsIds.sort();
-  @         output.innerHTML = permsIds.join('');
-  @       }
-  @     }
-  @   } catch (e) {
-  @     /* ignore errors */
-  @   }
-  @ }
-  @ </script>
   @ <table>
   @ <tr>
   @   <td class="usetupEditLabel">User ID:</td>
@@ -681,86 +645,60 @@ void user_edit(void){
 #define B(x) inherit[x]
   @ <table border=0><tr><td valign="top">
   if( g.perm.Setup ){
-    @  <label><input type="checkbox" name="as"%s(oa['s'])
-    @                onchange="updateCapabilityString()"/>
+    @  <label><input type="checkbox" name="as"%s(oa['s']) />
     @  Setup%s(B('s'))</label><br />
   }
-  @  <label><input type="checkbox" name="aa"%s(oa['a'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="aa"%s(oa['a']) />
   @  Admin%s(B('a'))</label><br />
-  @  <label><input type="checkbox" name="ad"%s(oa['d'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ad"%s(oa['d']) />
   @  Delete%s(B('d'))</label><br />
-  @  <label><input type="checkbox" name="ae"%s(oa['e'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ae"%s(oa['e']) />
   @  Email%s(B('e'))</label><br />
-  @  <label><input type="checkbox" name="ap"%s(oa['p'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ap"%s(oa['p']) />
   @  Password%s(B('p'))</label><br />
-  @  <label><input type="checkbox" name="ai"%s(oa['i'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ai"%s(oa['i']) />
   @  Check-In%s(B('i'))</label><br />
-  @  <label><input type="checkbox" name="ao"%s(oa['o'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ao"%s(oa['o']) />
   @  Check-Out%s(B('o'))</label><br />
-  @  <label><input type="checkbox" name="ah"%s(oa['h'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ah"%s(oa['h']) />
   @  Hyperlinks%s(B('h'))</label><br />
-  @  <label><input type="checkbox" name="ab"%s(oa['b'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ab"%s(oa['b']) />
   @  Attachments%s(B('b'))</label><br />
   @ </td><td><td width="40"></td><td valign="top">
-  @  <label><input type="checkbox" name="au"%s(oa['u'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="au"%s(oa['u']) />
   @  Reader%s(B('u'))</label><br />
-  @  <label><input type="checkbox" name="av"%s(oa['v'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="av"%s(oa['v']) />
   @  Developer%s(B('v'))</label><br />
-  @  <label><input type="checkbox" name="ag"%s(oa['g'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ag"%s(oa['g']) />
   @  Clone%s(B('g'))</label><br />
-  @  <label><input type="checkbox" name="aj"%s(oa['j'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="aj"%s(oa['j']) />
   @  Read Wiki%s(B('j'))</label><br />
-  @  <label><input type="checkbox" name="af"%s(oa['f'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="af"%s(oa['f']) />
   @  New Wiki%s(B('f'))</label><br />
-  @  <label><input type="checkbox" name="am"%s(oa['m'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="am"%s(oa['m']) />
   @  Append Wiki%s(B('m'))</label><br />
-  @  <label><input type="checkbox" name="ak"%s(oa['k'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ak"%s(oa['k']) />
   @  Write Wiki%s(B('k'))</label><br />
-  @  <label><input type="checkbox" name="al"%s(oa['l'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="al"%s(oa['l']) />
   @  Moderate Wiki%s(B('l'))</label><br />
   @ </td><td><td width="40"></td><td valign="top">
-  @  <label><input type="checkbox" name="ar"%s(oa['r'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ar"%s(oa['r']) />
   @  Read Ticket%s(B('r'))</label><br />
-  @  <label><input type="checkbox" name="an"%s(oa['n'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="an"%s(oa['n']) />
   @  New Tickets%s(B('n'))</label><br />
-  @  <label><input type="checkbox" name="ac"%s(oa['c'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ac"%s(oa['c']) />
   @  Append To Ticket%s(B('c'))</label><br />
-  @  <label><input type="checkbox" name="aw"%s(oa['w'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="aw"%s(oa['w']) />
   @  Write Tickets%s(B('w'))</label><br />
-  @  <label><input type="checkbox" name="aq"%s(oa['q'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="aq"%s(oa['q']) />
   @  Moderate Tickets%s(B('q'))</label><br />
-  @  <label><input type="checkbox" name="at"%s(oa['t'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="at"%s(oa['t']) />
   @  Ticket Report%s(B('t'))</label><br />
-  @  <label><input type="checkbox" name="ax"%s(oa['x'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ax"%s(oa['x']) />
   @  Private%s(B('x'))</label><br />
-  @  <label><input type="checkbox" name="ay"%s(oa['y'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="ay"%s(oa['y']) />
   @  Write Unversioned%s(B('y'))</label><br />
-  @  <label><input type="checkbox" name="az"%s(oa['z'])
-  @                onchange="updateCapabilityString()" />
+  @  <label><input type="checkbox" name="az"%s(oa['z']) />
   @  Download Zip%s(B('z'))</label>
   @ </td></tr>
   @ </table>
@@ -804,7 +742,7 @@ void user_edit(void){
   @ </table>
   @ </div></form>
   @ </div>
-  @ <script>updateCapabilityString();</script>
+  style_load_one_js_file("useredit.js");
   @ <h2>Privileges And Capabilities:</h2>
   @ <ul>
   if( higherUser ){
@@ -1225,17 +1163,19 @@ void setup_access(void){
   onoff_attribute(
       "Enable hyperlinks for \"nobody\" based on User-Agent and Javascript",
       "auto-hyperlink", "autohyperlink", 1, 0);
-  @ <p>Enable hyperlinks (the equivalent of the "h" permission) for all users
-  @ including user "nobody", as long as (1) the User-Agent string in the
+  @ <p>Enable hyperlinks (the equivalent of the "h" permission) for all users,
+  @ including user "nobody", as long as
+  @ <ol><li>the User-Agent string in the
   @ HTTP header indicates that the request is coming from an actual human
-  @ being and not a robot or spider and (2) the user agent is able to
-  @ run Javascript in order to set the href= attribute of hyperlinks.  Bots
-  @ and spiders can forge a User-Agent string that makes them seem to be a
-  @ normal browser and they can run javascript just like browsers.  But most
-  @ bots do not go to that much trouble so this is normally an effective
-  @ defense.<p>
+  @ being, and
+  @ <li>the user agent is able to
+  @ run Javascript in order to set the href= attribute of hyperlinks, and
+  @ <li>mouse movement is detected (optional - see the checkbox below), and
+  @ <li>a number of milliseconds have passed since the page loaded.</ol>
   @
-  @ <p>You do not normally want a bot to walk your entire repository because
+  @ <p>This setting is designed to give easy access to humans while
+  @ keeping out robots and spiders.
+  @ You do not normally want a robot to walk your entire repository because
   @ if it does, your server will end up computing diffs and annotations for
   @ every historical version of every file and creating ZIPs and tarballs of
   @ every historical check-in, which can use a lot of CPU and bandwidth
@@ -1243,20 +1183,15 @@ void setup_access(void){
   @
   @ <p>Additional parameters that control this behavior:</p>
   @ <blockquote>
-  onoff_attribute("Enable hyperlinks for humans as deduced from the UserAgent "
-                  "string", "auto-hyperlink-ishuman", "ahis", 0, 0);
-  @ <br />
   onoff_attribute("Require mouse movement before enabling hyperlinks",
                   "auto-hyperlink-mouseover", "ahmo", 0, 0);
   @ <br />
   entry_attribute("Delay in milliseconds before enabling hyperlinks", 5,
-                  "auto-hyperlink-delay", "ah-delay", "10", 0);
+                  "auto-hyperlink-delay", "ah-delay", "50", 0);
   @ </blockquote>
-  @ <p>Hyperlinks for user "nobody" are normally enabled as soon as the page
-  @ finishes loading.  But the first check-box below can be set to require mouse
-  @ movement before enabling the links. One can also set a delay prior to enabling
-  @ links by enter a positive number of milliseconds in the entry box above.</p>
-  @ (Properties: "auto-hyperlink", "auto-hyperlink-ishuman",
+  @ <p>For maximum robot defense, the "require mouse movement" should
+  @ be turned on and the "Delay" should be at least 50 milliseconds.</p>
+  @ (Properties: "auto-hyperlink",
   @ "auto-hyperlink-mouseover", and "auto-hyperlink-delay")</p>
 
   @ <hr />
@@ -1415,7 +1350,7 @@ void setup_login_group(void){
     @ <hr /><h2>Implementation Details</h2>
     @ <p>The following are fields from the CONFIG table related to login-groups,
     @ provided here for instructional and debugging purposes:</p>
-    @ <table border='1' id='configTab'>
+    @ <table border='1' class='sortable' data-column-types='ttt' data-init-sort='1'>
     @ <thead><tr>
     @ <th>Config.Name<th>Config.Value<th>Config.mtime</tr>
     @ </thead><tbody>
@@ -1431,7 +1366,7 @@ void setup_login_group(void){
     }
     db_finalize(&q);
     @ </tbody></table>
-    output_table_sorting_javascript("configTab","ttt",1);
+    style_table_sorter();
   }
   style_footer();
 }
@@ -1462,8 +1397,8 @@ void setup_timeline(void){
   db_begin_transaction();
   @ <form action="%s(g.zTop)/setup_timeline" method="post"><div>
   login_insert_csrf_secret();
-
   @ <p><input type="submit"  name="submit" value="Apply Changes" /></p>
+
   @ <hr />
   onoff_attribute("Allow block-markup in timeline",
                   "timeline-block-markup", "tbm", 0, 0);
@@ -1655,12 +1590,22 @@ void setup_config(void){
   @ engines as well as a short RSS description.
   @ (Property: "project-description")</p>
   @ <hr />
-  entry_attribute("Tarball and ZIP-archive Prefix", 20, "short-project-name", "spn", "", 0);
+  entry_attribute("Tarball and ZIP-archive Prefix", 20, "short-project-name",
+                  "spn", "", 0);
   @ <p>This is used as a prefix on the names of generated tarballs and ZIP archive.
   @ For best results, keep this prefix brief and avoid special characters such
   @ as "/" and "\".
   @ If no tarball prefix is specified, then the full Project Name above is used.
   @ (Property: "short-project-name")
+  @ </p>
+  @ <hr />
+  entry_attribute("Download Tag", 20, "download-tag", "dlt", "trunk", 0);
+  @ <p>The <a href='%R/download'>/download</a> page is designed to provide 
+  @ a convenient place for newbies
+  @ to download a ZIP archive or a tarball of the project.  By default, the latest
+  @ trunk check-in is downloaded.  Change this tag to something else (ex: release)
+  @ to alter the behavior of the /download page.
+  @ (Property: "download-tag")
   @ </p>
   @ <hr />
   onoff_attribute("Enable WYSIWYG Wiki Editing",
@@ -1703,7 +1648,7 @@ void setup_config(void){
   @ No sanitization is done. This means that it is very possible for malicious
   @ users to inject dangerous HTML, CSS and JavaScript code into your wiki.</p>
   @ <p>This should <strong>only</strong> be enabled when wiki editing is limited
-  @ to trusted users. It should <strong>not</strong> be used on a publically
+  @ to trusted users. It should <strong>not</strong> be used on a publicly
   @ editable wiki.</p>
   @ (Property: "wiki-use-html")
   @ <hr />
@@ -1773,7 +1718,7 @@ void setup_adunit(void){
     return;
   }
   db_begin_transaction();
-  if( P("clear")!=0 ){
+  if( P("clear")!=0 && cgi_csrf_safe(1) ){
     db_multi_exec("DELETE FROM config WHERE name GLOB 'adunit*'");
     cgi_replace_parameter("adunit","");
   }
@@ -1862,7 +1807,9 @@ void setup_logo(void){
     return;
   }
   db_begin_transaction();
-  if( P("setlogo")!=0 && zLogoMime && zLogoMime[0] && szLogoImg>0 ){
+  if( !cgi_csrf_safe(1) ){
+    /* Allow no state changes if not safe from CSRF */
+  }else if( P("setlogo")!=0 && zLogoMime && zLogoMime[0] && szLogoImg>0 ){
     Blob img;
     Stmt ins;
     blob_init(&img, aLogoImg, szLogoImg);
@@ -1997,7 +1944,7 @@ int raw_sql_query_authorizer(
 ** Requires Admin privileges.
 */
 void sql_page(void){
-  const char *zQ = P("q");
+  const char *zQ;
   int go = P("go")!=0;
   login_check_credentials();
   if( !g.perm.Setup ){
@@ -2006,6 +1953,7 @@ void sql_page(void){
   }
   add_content_sql_commands(g.db);
   db_begin_transaction();
+  zQ = cgi_csrf_safe(1) ? P("q") : 0;
   style_header("Raw SQL Commands");
   @ <p><b>Caution:</b> There are no restrictions on the SQL that can be
   @ run by this page.  You can do serious and irrepairable damage to the
@@ -2205,8 +2153,9 @@ void page_admin_log(){
     "SELECT datetime(time,'unixepoch'), who, page, what "
     "FROM admin_log "
     "ORDER BY time DESC");
-
-  @ <table id="adminLogTable" class="adminLogTable" width="100%%">
+  style_table_sorter();
+  @ <table class="sortable adminLogTable" width="100%%" \
+  @  data-column-types='Tttx' data-init-sort='1'>
   @ <thead>
   @ <th>Time</th>
   @ <th>User</th>
@@ -2232,8 +2181,6 @@ void page_admin_log(){
   if( counter>ofst+limit ){
     @ <p><a href="admin_log?n=%d(limit)&x=%d(limit+ofst)">[Older]</a></p>
   }
-
-  output_table_sorting_javascript("adminLogTable", "Tttx", 1);
   style_footer();
 }
 
@@ -2328,6 +2275,7 @@ static void setup_update_url_alias(
   const char *zNewName,
   const char *zValue
 ){
+  if( !cgi_csrf_safe(1) ) return;
   if( zNewName[0]==0 || zValue[0]==0 ){
     if( zOldName[0] ){
       blob_append_sql(pSql,
